lyd_validate_obsolete: change log level warning -> debug
A library should not log directly to syslog, this is the responsibility
of the application.  In our case sysrepo, with a knob to disable it when
you know what you're doing or have other ways of handling deprecated and
obsolete nodes — e.g., automatically migrating them to a new tree.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>